requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
elevenlabs==2.11.0
plaid-python==21.0.0
